- Added intermediate component, workspace, to hold Lists and Todos component. Still need to work on integrating those two components together.  - Added feature to delete list.  - Minor CSS changes.  - Linted App.js

+++ src/App.js
@@ -2,16 +2,15 @@ import React, { Component } from "react"
 import "./App.css"
 import firebase from "./services/Firebase/firebase"
 import firebaseUiConfig from "./services/Firebase/firebaseUiConf"
-import db from './services/Firebase/firestore'
+import db from "./services/Firebase/firestore"
 
 import { Router, navigate } from "@reach/router"
 import { v1 as uuid } from "uuid"
 
 import Home from "./components/Home/Home"
 import Login from "./components/Login/Login"
-import Todos from "./components/Todos/Todos.js"
 import Layout from "./components/Layout/Layout"
-import Lists from './components/Lists/Lists'
+import Workspace from "./components/Workspace/Workspace"
 
 class App extends Component {
 	state = {
@@ -32,7 +31,7 @@ class App extends Component {
 				}
 				this.setState({ user: currentUser })
 
-				let todosRef = db.collection('todos').doc(`${this.state.user.uid}`)
+				let todosRef = db.collection("todos").doc(`${this.state.user.uid}`)
 				todosRef.onSnapshot(snapShot => {
 					let data = snapShot.data()
 					let todos = []
@@ -44,7 +43,7 @@ class App extends Component {
 					}
 				})
 
-				let listsRef = db.collection('lists').doc(`${this.state.user.uid}`)
+				let listsRef = db.collection("lists").doc(`${this.state.user.uid}`)
 				listsRef.onSnapshot(snapShot => {
 					let data = snapShot.data()
 					let lists = []
@@ -56,28 +55,34 @@ class App extends Component {
 					}
 				})
 
-				//navigate("/todos")
+				navigate("/workspace")
 			}
 		})
 	}
 
 	toggleTodoCompletion = (evt, todoId) => {
-		let todo = this.state.todos.filter(todo => todo.id === todoId ? todo : false)
+		let todo = this.state.todos.filter(todo => (todo.id === todoId ? todo : false))
 		todo[0].done = !todo[0].done
-		let ref = db.collection('todos').doc(`${this.state.user.uid}`)
-		ref.set({
-			[todoId]: todo[0]
-		}, { merge: true })
+		let ref = db.collection("todos").doc(`${this.state.user.uid}`)
+		ref.set(
+			{
+				[todoId]: todo[0]
+			},
+			{ merge: true }
+		)
 	}
 
 	editTodoDescription = (todoId, description) => {
-		let todo = this.state.todos.filter(todo => todo.id === todoId ? todo : false)
+		let todo = this.state.todos.filter(todo => (todo.id === todoId ? todo : false))
 		todo[0].description = description
 
-		let ref = db.collection('todos').doc(`${this.state.user.uid}`)
-		ref.set({
-			[todoId]: todo[0]
-		}, { merge: true })
+		let ref = db.collection("todos").doc(`${this.state.user.uid}`)
+		ref.set(
+			{
+				[todoId]: todo[0]
+			},
+			{ merge: true }
+		)
 	}
 
 	addTodo = description => {
@@ -87,14 +92,17 @@ class App extends Component {
 			done: false
 		}
 
-		let ref = db.collection('todos').doc(`${this.state.user.uid}`)
-		ref.set({
-			[todo.id]: todo
-		}, { merge: true })
+		let ref = db.collection("todos").doc(`${this.state.user.uid}`)
+		ref.set(
+			{
+				[todo.id]: todo
+			},
+			{ merge: true }
+		)
 	}
 
 	deleteTodo = todoId => {
-		let ref = db.collection('todos').doc(`${this.state.user.uid}`)
+		let ref = db.collection("todos").doc(`${this.state.user.uid}`)
 		ref.update({
 			[todoId]: firebase.firestore.FieldValue.delete()
 		})
@@ -107,17 +115,20 @@ class App extends Component {
 	}
 
 	editListName = (listId, listName) => {
-		let list = this.state.lists.filter(list => list.id === listId ? list : false)
+		let list = this.state.lists.filter(list => (list.id === listId ? list : false))
 		list[0].name = listName
 
-		let ref = db.collection('lists').doc(`${this.state.user.uid}`)
-		ref.set({
-			[listId]: list[0]
-		}, { merge: true })
+		let ref = db.collection("lists").doc(`${this.state.user.uid}`)
+		ref.set(
+			{
+				[listId]: list[0]
+			},
+			{ merge: true }
+		)
 	}
 
 	addList = evt => {
-		if (evt.target.value === '') {
+		if (evt.target.value === "") {
 			return
 		}
 
@@ -125,19 +136,26 @@ class App extends Component {
 			id: uuid(),
 			name: evt.target.value
 		}
-		let lists = [...this.state.lists, list]
 
-		let ref = db.collection('lists').doc(`${this.state.user.uid}`)
-		ref.set({
-			[list.id]: list
-		}, { merge: true })
+		let ref = db.collection("lists").doc(`${this.state.user.uid}`)
+		ref.set(
+			{
+				[list.id]: list
+			},
+			{ merge: true }
+		)
 	}
 
 	deleteList = listId => {
-		let ref = db.collection('lists').doc(`${this.state.user.uid}`)
-		ref.update({
-			[listId]: firebase.firestore.FieldValue.delete()
-		})
+		let doDeleteList = window.confirm("Are you sure you want to delete this list? \n\nThis action cannot be undone.")
+
+		if (doDeleteList) {
+			let ref = db.collection("lists").doc(`${this.state.user.uid}`)
+			ref.update({
+				[listId]: firebase.firestore.FieldValue.delete()
+			})
+		}
+		return
 	}
 
 	render() {
@@ -146,22 +164,19 @@ class App extends Component {
 				<Router className="d-flex flex-column flex-grow-1">
 					<Home path="/" />
 					<Login path="/login" firebaseUiConfig={firebaseUiConfig} firebaseAuth={firebase.auth} />
-					<Todos
-						path="/todos"
+					<Workspace
+						path="/workspace"
 						todos={this.state.todos}
 						toggleTodoCompletion={this.toggleTodoCompletion}
 						editTodoDescription={this.editTodoDescription}
 						deleteTodo={this.deleteTodo}
 						addTodo={this.addTodo}
 						user={this.state.user}
-					/>
-					<Lists
-						path='/lists'
 						lists={this.state.lists}
 						addList={this.addList}
 						editListName={this.editListName}
 						deleteList={this.deleteList}
-					></Lists>
+					/>
 				</Router>
 			</Layout>
 		)
